requirements.txt
# LangChain stack (0.3 line)
langchain-core==0.3.66
langchain==0.3.26
langchain-community==0.3.26
langchain-ibm==0.3.13
ibm-watsonx-ai==1.3.26

# NumPy wheel that already supports cp312
numpy==1.26.4          # last 1.x release; wheels for 3.12 exist

# UI / voice layer
streamlit==1.34.0
SpeechRecognition==3.14.3
pydub==0.25.1

python-dotenv==1.0.1

